Added class and interface of Service, ServiceImpl and UserDAO and Carpet ValidateUser

+++ src/main/java/com/eam/LevelUpCorp/persistenceLayer/entity/FileResourceEntity.java
@@ -15,11 +15,11 @@ public class FileResourceEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
     private String fileName;
     private String fileType;
-    private int moduleId;
-    private int evaluationId;
+    private Long moduleId;
+    private Long evaluationId;
 
 
 }
